ml/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
onnx>=1.15.0
onnxruntime>=1.17.0
Pillow>=10.0.0
requests>=2.31.0
tqdm>=4.66.0
scikit-learn>=1.3.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.13.0
pandas>=2.0.0
tensorboard>=2.14.0
nvidia-dali-cuda120
